+++ package.js
@@ -19,10 +19,14 @@ Package.onTest( function( api ){
 });
 
 function configure( api ){
+    const _use = function(){
+        api.use( ...arguments );
+        api.imply( ...arguments );
+    };
     api.versionsFrom([ '2.9.0', '3.0-rc.0' ]);
-    api.use( 'aldeed:schema-deny@3.0.0 || 4.0.0' );
-    api.use( 'check' );
-    api.use( 'ecmascript' );
-    api.use( 'matb33:collection-hooks@1.3.0 || 2.0.0-beta.0' );
-    api.use( 'pwix:collection-behaviours@2.0.0' );
+    _use( 'aldeed:schema-deny@3.0.0 || 4.0.0' );
+    _use( 'check' );
+    _use( 'ecmascript' );
+    _use( 'matb33:collection-hooks@1.3.0 || 2.0.0-beta.0' );
+    _use( 'pwix:collection-behaviours@2.0.0' );
 }