+++ package.js
@@ -25,6 +25,7 @@ function configure( api ){
     };
     api.versionsFrom([ '2.9.0', '3.0-rc.0' ]);
     _use( 'aldeed:schema-deny@3.0.0 || 4.0.0' );
+    _use( 'aldeed:simple-schema@1.13.1' );
     _use( 'check' );
     _use( 'ecmascript' );
     _use( 'matb33:collection-hooks@1.3.0 || 2.0.0-beta.0' );