+++ src/common/js/constants.js
@@ -1,0 +1,13 @@
+/*
+ * pwix:collection-timestampable/src/common/js/constants.js
+ */
+
+Timestampable.C = {
+
+    // verbosity levels
+    Verbose: {
+        NONE: 0,
+        CONFIGURE:      0x01 <<  0,
+        FUNCTIONS:      0x01 <<  1
+    }
+};