+++ src/common/js/global.js
@@ -1,0 +1,6 @@
+/*
+ * pwix:collection-timestampable/src/common/js/global.js
+ */
+
+Timestampable = {
+};