+++ src/common/js/index.js
@@ -2,4 +2,9 @@
  * pwix:collection-timestampable/src/common/js/index.js
  */
 
+import './global.js';
+import './constants.js';
+import './trace.js';
+//
+import './configure.js';
 import './timestampable.js';